requirements.txt
scipy==1.4.1
numpy==1.18.1
tensorflow_gpu==2.2.0
matplotlib==3.1.3
Keras==2.3.1
Pillow==7.1.2
scikit_learn==0.23.1
tensorflow==2.2.0
